refactor: address comments and rollback to the initial component
